+++ src/settings.ts
@@ -1,11 +1,15 @@
+import * as _ from 'lodash';
 import {TemplateSrv} from "grafana/app/features/templating/template_srv";
 
 export default class SmartSettings {
 
     private _request: Object;
+    private request_variables: Object;
 
     constructor(private model: Object, private templateSrv: TemplateSrv, private request?: Object) {
         this._request = this.request || {};
+        this.request_variables = {};
+        this.loadVariablesFromRequest();
         this.establishProperties();
     }
 
@@ -18,10 +22,7 @@ export default class SmartSettings {
         }
     }
 
-    interpolateVariable(name, variables?: Object) {
-
-        variables = variables || {};
-
+    loadVariablesFromRequest() {
         // Put all request parameters from url into the variable
         // interpolation dictionary, prefixed by `request_`.
         //
@@ -33,11 +34,22 @@ export default class SmartSettings {
         for (let key in this._request) {
             const value = this._request[key];
             key = 'request_' + key;
-            variables[key] = value;
+            this.request_variables[key] = value;
         }
+    }
+
+    interpolateVariable(name, variables?: Object) {
+
+        variables = variables || {};
+
+        variables = _.cloneDeep(variables);
+        _.merge(variables, this.request_variables);
 
         // By default, use vanilla control option attribute from panel data model.
         let value = this.model[name];
+
+        // Optionally, use `panel-` variables from request query parameters.
+        value = this.getVariableFromRequest(name, value);
 
         // Interpolate the dashboard and dataPoint variables.
         return this.interpolateVariableValue(value, variables);
@@ -60,4 +72,36 @@ export default class SmartSettings {
         return scopedVars;
     }
 
+    getVariableFromRequest(name, value) {
+
+        // When given, use request variable "panel-*", making things like these possible.
+        // - ?panel-mapCenterLatitude=62.2
+        // - ?panel-showZoomControl=false
+        // - ?panel-clickthroughURL=/path/to/?geohash=$point_geohash
+        // - https://daq.example.org/d/D1Fx12kWk/magic-dashboard?panel-clickthroughURL=/path/to/?foobar=$request_foobar&foobar=hello
+        const panel_query_name = 'panel-' + name;
+        const panel_query_value = this._request[panel_query_name];
+        if (panel_query_value !== undefined) {
+
+            // Apply appropriate type conversion. This is important for booleans.
+            if (typeof value == 'boolean') {
+                value = asBool(panel_query_value);
+            } else {
+                value = panel_query_value;
+            }
+        }
+
+        return value;
+
+    }
+
+}
+
+function asBool(value) {
+    // https://stackoverflow.com/questions/263965/how-can-i-convert-a-string-to-boolean-in-javascript/1414175#1414175
+    switch(value.toLowerCase().trim()){
+        case "true": case "yes": case "1": return true;
+        case "false": case "no": case "0": case null: return false;
+        default: return Boolean(value);
+    }
 }